third_party/WebKit/LayoutTests/fast/dom/Element/id-in-insert-hr-expected.txt

Test using id in execCommand(InsertHorizontalRule)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS hr is non-null.
PASS successfullyParsed is true

TEST COMPLETE

foobar
